backend-python/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
pydantic==2.4.0
python-dotenv==1.0.0
httpx==0.25.0
APScheduler==3.10.4
